requirements.txt
google-api-python-client
python-dotenv
pandas
matplotlib
seaborn
koreanize-matplotlib
requests